ml-model/requirements.txt
tensorflow>=2.20.0
numpy>=1.26.0
pandas>=2.1.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0
pillow>=10.0.0
opencv-python>=4.8.0.76
tqdm>=4.66.1
jupyter>=1.0.0
python-dotenv==1.0.0

